Add in the ability to have arbitrary arguments. (#7)
That is, add the ability to have arbitrary colcon arguments.

Hidden in this refactor is a change in the default arguments
used for packages.  That is, if packages are specified, the
default is now to use --packages-above-and-dependencies for
builds and --packages-above for tests.  If --only-fixes-test
is specified (a new flag), the old behavior of using
--packages-up-to for build and --packages-select for test is
used.  Arguably this reflects our current best practice in
ROS 2, where we want to make sure all packages that depend on
this change are tested as well.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>